requirements.txt
neo4j>=5.5.0
python-dotenv>=0.21.0
openai>=0.27.0
anthropic>=0.4.0
flask>=2.0.0 